+++ backend/tasks_lib/llm_lib/llm_ops.py
@@ -3,12 +3,14 @@ from time import sleep
 from typing import Generator
 from langchain_openai import ChatOpenAI
 from langchain_core.prompts import PromptTemplate
-from langchain_core.runnables import RunnablePassthrough
+from langchain_core.runnables import RunnablePassthrough, RunnableLambda
 from langchain_core.output_parsers import StrOutputParser
+from langchain_core.prompt_values import StringPromptValue
 import requests
 from pydantic import SecretStr
 from tasks_lib.cmd_line_opts import IS_DUMMY_LLM
 from .dummy import DummyLLM
+from common import log
 from common.helpers import chunks
 from common.parsed_url import ParsedUrl
 from common.enums.gllms_types import GLLMsTypes
@@ -39,6 +41,7 @@ class LLMOps:
         self.temperature = 0.0
         self.llm_max_tokens = 500
         self.context_json: str = ''
+        self.sent_to_llm: str = ''
         self.answer: str = ''
 
     def check_working(self) -> bool:
@@ -67,17 +70,29 @@ class LLMOps:
         # Combine all context
         return "\n\n".join(context_parts)
     
+    def _stream_dummy_answer(self) -> Generator[str, None, None]:
+        dummy_answer = f"Dummy answer to query:\n{self.query_text}\n{DummyLLM.fake_answer}"
+        self.answer = ""
+        for chunk in chunks(list(dummy_answer), 50):
+            chunk_str = ''.join(chunk)
+            self.answer += chunk_str
+            yield chunk_str
+            sleep(1)
+        return
+    
     def stream_to_llm(self) -> Generator[str, None, None]:
         # Initialize LLM
         if IS_DUMMY_LLM or (self.llm_type == GLLMsTypes.DUMMY):
-            dummy_answer = f"Dummy answer to query:\n{self.query_text}\n{DummyLLM.fake_answer}"
-            self.answer = ""
-            for chunk in chunks(list(dummy_answer), 50):
-                chunk_str = ''.join(chunk)
-                self.answer += chunk_str
-                yield chunk_str
-                sleep(1)
+            yield from self._stream_dummy_answer()
             return
+        
+        def capture_prompt(prompt_value):
+            if isinstance(prompt_value, StringPromptValue):
+                self.sent_to_llm = str(prompt_value.text)
+            else:
+                log.error(f"Prompt value is not StringPromptValue, but {type(prompt_value)}")
+            return prompt_value
+
         full_context = self.prepare_context()
         llm = ChatOpenAI(
             model=self.llm_model,
@@ -94,10 +109,10 @@ class LLMOps:
         rag_chain = (
             {"context": lambda x: full_context, "question": RunnablePassthrough()}
             | prompt
+            | RunnableLambda(capture_prompt) 
             | llm
             | StrOutputParser()
         )
-        
         # Stream the response
         for chunk in rag_chain.stream(self.query_text):
             self.answer += chunk